+++ hera/main.py
@@ -312,7 +312,7 @@ PICKY_FLAGS = {
     "--stdout": ["assemble"],
 }
 
-VERSION = "hera-py 0.6.0 for HERA version 2.4"
+VERSION = "hera-py 0.7.0 for HERA version 2.4"
 HELP = """\
 hera: an interpreter for the Haverford Educational RISC Architecture.