+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 **NOTE**: As permitted by semantic versioning, backward compatibility is NOT maintained for initial development, i.e. releases before 1.0.0.
 
 ## Unreleased
+Nothing yet.
+
+## [0.7.0] - 2019-02-13
 ### Added
 - The `assemble` subcommand, for assembling HERA programs into raw machine code.
 - The `clear` debugging command, for clearing breakpoints.